refactor: move header & footer to layouts folder
