+++ apps/webapp/src/pages/index.jsx
@@ -20,10 +20,10 @@ import { useDispatch } from 'react-redux';
 import { v4 as uuidv4 } from 'uuid';
 import { ImagePreview } from '@/components/elements/ImagePreview/ImagePreview';
 import { StepProgress } from '@/components/elements/StepProgress';
+import { Footer } from '@/components/layouts/Footer';
+import { Header } from '@/components/layouts/Header';
 import { DownloadMenu } from '@/components/modules/DownloadMenu';
 import { Faq } from '@/components/modules/Faq';
-import { Footer } from '@/components/modules/Footer';
-import { Header } from '@/components/modules/Header';
 import { SubmitForm } from '@/components/modules/SubmitForm';
 import { MAX_FILES } from '@/constants/constants.js';
 import { socket } from '@/socket/socket.js';
